Consolidate inline commands
Now with some new base classes.
