+++ src/nl/hannahsten/texifyidea/refactoring/inlinecommand/LatexInlineCommandDialog.kt
@@ -3,12 +3,7 @@ package nl.hannahsten.texifyidea.refactoring.inlinecommand
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.NlsContexts.DialogTitle
 import com.intellij.psi.PsiElement
-import com.intellij.psi.PsiNameIdentifierOwner
 import com.intellij.psi.search.GlobalSearchScope
-import com.intellij.psi.search.searches.ReferencesSearch
-import com.intellij.refactoring.JavaRefactoringSettings
-import com.intellij.refactoring.RefactoringBundle
-import com.intellij.refactoring.inline.InlineOptionsDialog
 import nl.hannahsten.texifyidea.psi.LatexCommands
 import nl.hannahsten.texifyidea.util.definitionCommand
 
@@ -23,39 +18,18 @@ class LatexInlineCommandDialog(
     project: Project?,
     private val myDefinition: LatexCommands,
     private val myReference: PsiElement?,
-    myInvokedOnReference: Boolean,
+    invokedOnReference: Boolean,
 ) :
-    InlineOptionsDialog(project, true, myDefinition) {
-
-    val myOccurrencesNumber: Int
+    LatexInlineDialog(project, myDefinition, invokedOnReference) {
 
     init {
-        super.myInvokedOnReference = myInvokedOnReference
-
-        myOccurrencesNumber = getNumberOfOccurrences(myDefinition.definitionCommand())
         title = refactoringName
         init()
-    }
-
-    override fun getNumberOfOccurrences(nameIdentifierOwner: PsiNameIdentifierOwner?): Int {
-        val tempreferences = ReferencesSearch.search(nameIdentifierOwner as PsiElement).findAll().asSequence()
-
-        return tempreferences
-            .distinct()
-            .toList().size
-    }
-
-    override fun allowInlineAll(): Boolean {
-        return true
     }
 
     override fun getNameLabelText(): String {
         return if (myOccurrencesNumber > -1) "Command " + myDefinition.name + " has " + myOccurrencesNumber + " ocurrences"
         else "Command " + myDefinition.name
-    }
-
-    override fun getBorderTitle(): String {
-        return RefactoringBundle.message("inline.method.border.title")
     }
 
     override fun getInlineThisText(): String {
@@ -70,36 +44,22 @@ class LatexInlineCommandDialog(
         return if (myDefinition.isWritable) "Inline all and keep the command" else super.getKeepTheDeclarationText()
     }
 
-    public override fun doAction() {
+    override fun doAction() {
         invokeRefactoring(
             LatexInlineCommandProcessor(
                 project, myDefinition, myReference, isInlineThisOnly, isKeepTheDeclaration, GlobalSearchScope.projectScope(myProject)
             )
         )
-        val settings = JavaRefactoringSettings.getInstance()
-        if (myRbInlineThisOnly.isEnabled && myRbInlineAll.isEnabled) {
-            settings.INLINE_METHOD_THIS = isInlineThisOnly
-        }
-        if (myKeepTheDeclaration != null && myKeepTheDeclaration!!.isEnabled) {
-            settings.INLINE_METHOD_KEEP = isKeepTheDeclaration
-        }
+        updateSettingsPreferences()
     }
 
-    override fun isInlineThis(): Boolean {
-        return JavaRefactoringSettings.getInstance().INLINE_METHOD_THIS
-    }
-
-    override fun isKeepTheDeclarationByDefault(): Boolean {
-        return JavaRefactoringSettings.getInstance().INLINE_METHOD_KEEP
+    override fun getNumberOfOccurrences(): Int {
+        return super.getNumberOfOccurrences(myDefinition.definitionCommand())
     }
 
     companion object {
 
         val refactoringName: @DialogTitle String
             get() = "Inline Command"
-    }
-
-    override fun hasHelpAction(): Boolean {
-        return false
     }
 }
