+++ src/nl/hannahsten/texifyidea/refactoring/inlinefile/LatexInlineFileHandler.kt
@@ -1,7 +1,5 @@
 package nl.hannahsten.texifyidea.refactoring.inlinefile
 
-import com.intellij.lang.Language
-import com.intellij.lang.refactoring.InlineActionHandler
 import com.intellij.notification.Notification
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.application.ApplicationManager
@@ -10,9 +8,9 @@ import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.PsiUtilBase.getElementAtCaret
-import nl.hannahsten.texifyidea.LatexLanguage
 import nl.hannahsten.texifyidea.file.LatexFile
 import nl.hannahsten.texifyidea.psi.LatexCommands
+import nl.hannahsten.texifyidea.refactoring.inlinecommand.LatexInlineHandler
 import nl.hannahsten.texifyidea.reference.InputFileReference
 import nl.hannahsten.texifyidea.util.Log
 import nl.hannahsten.texifyidea.util.files.isLatexFile
@@ -23,11 +21,7 @@ import nl.hannahsten.texifyidea.util.firstParentOfType
  *
  * @author jojo2357
  */
-class LatexInlineFileHandler : InlineActionHandler() {
-
-    override fun isEnabledForLanguage(language: Language?): Boolean {
-        return language == LatexLanguage
-    }
+class LatexInlineFileHandler : LatexInlineHandler() {
 
     override fun canInlineElement(element: PsiElement?): Boolean {
         return canInlineLatexElement(element)
@@ -66,53 +60,56 @@ class LatexInlineFileHandler : InlineActionHandler() {
             ).notify(project)
         }
     }
-}
 
-/**
- * This is static only so that the unit tests can use it, and also since it can be static
- */
-fun canInlineLatexElement(element: PsiElement?): Boolean {
-    return when (element) {
-        is LatexFile -> {
-            element.containingFile.isLatexFile()
+    companion object {
+
+        /**
+         * This is static only so that the unit tests can use it, and also since it can be static
+         */
+        fun canInlineLatexElement(element: PsiElement?): Boolean {
+            return when (element) {
+                is LatexFile -> {
+                    element.containingFile.isLatexFile()
+                }
+
+                is LatexCommands -> {
+                    (element.references.filterIsInstance<InputFileReference>().isNotEmpty())
+                }
+
+                else -> false
+            }
         }
 
-        is LatexCommands -> {
-            (element.references.filterIsInstance<InputFileReference>().isNotEmpty())
+        fun resolveInlineFile(element: PsiElement) = when (element) {
+            // If we tried to inline the file in file view, or the argument in `input`
+            is LatexFile -> element
+            // If the caret was on the `input` command
+            is LatexCommands -> {
+                element.references.filterIsInstance<InputFileReference>()[0].resolve(false) as? LatexFile
+            }
+
+            else -> null
         }
 
-        else -> false
-    }
-}
-
-fun resolveInlineFile(element: PsiElement) = when (element) {
-    // If we tried to inline the file in file view, or the argument in `input`
-    is LatexFile -> element
-    // If the caret was on the `input` command
-    is LatexCommands -> {
-        element.references.filterIsInstance<InputFileReference>()[0].resolve(false) as? LatexFile
-    }
-
-    else -> null
-}
-
-/**
- * Resolves the command that the inlining was called on. If called from file view then expected return is null
- */
-fun getReference(
-    element: PsiElement,
-    editor: Editor?
-): PsiElement? {
-    return if (element is LatexCommands)
-        element
-    else if (editor != null) {
-        val ref: LatexCommands? = getElementAtCaret(editor)?.firstParentOfType(LatexCommands::class)
-        if (ref != null)
-            ref
-        else {
-            Log.warn("Could not find the command element for " + element.text)
-            null
+        /**
+         * Resolves the command that the inlining was called on. If called from file view then expected return is null
+         */
+        fun getReference(
+            element: PsiElement,
+            editor: Editor?
+        ): PsiElement? {
+            return if (element is LatexCommands)
+                element
+            else if (editor != null) {
+                val ref: LatexCommands? = getElementAtCaret(editor)?.firstParentOfType(LatexCommands::class)
+                if (ref != null)
+                    ref
+                else {
+                    Log.warn("Could not find the command element for " + element.text)
+                    null
+                }
+            }
+            else null
         }
     }
-    else null
 }
