+++ test/nl/hannahsten/texifyidea/refactoring/InlineFileTest.kt
@@ -11,10 +11,10 @@ import com.intellij.refactoring.inline.InlineOptions
 import com.intellij.testFramework.LightPlatformCodeInsightTestCase
 import junit.framework.TestCase
 import nl.hannahsten.texifyidea.file.LatexFileType
+import nl.hannahsten.texifyidea.refactoring.inlinecommand.LatexInlineCommandHandler.Companion.getReference
+import nl.hannahsten.texifyidea.refactoring.inlinefile.LatexInlineFileHandler.Companion.canInlineLatexElement
+import nl.hannahsten.texifyidea.refactoring.inlinefile.LatexInlineFileHandler.Companion.resolveInlineFile
 import nl.hannahsten.texifyidea.refactoring.inlinefile.LatexInlineFileProcessor
-import nl.hannahsten.texifyidea.refactoring.inlinefile.canInlineLatexElement
-import nl.hannahsten.texifyidea.refactoring.inlinefile.getReference
-import nl.hannahsten.texifyidea.refactoring.inlinefile.resolveInlineFile
 import nl.hannahsten.texifyidea.util.runWriteAction
 import org.jetbrains.annotations.NonNls
 import java.io.File
